Create file hierarchy
