+++ QtPiDeck/include/appStartup.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace QtPiDeck {
+void registerQmlTypes();
+void setCursorVisibility();
+}
