+++ QtPiDeck/qml.qrc
@@ -1,5 +1,5 @@
 <RCC>
     <qresource prefix="/">
-        <file>main.qml</file>
+        <file>qml/pages/main.qml</file>
     </qresource>
 </RCC>
