+++ QtPiDeck/src/main.cpp
@@ -1,10 +1,8 @@
 #include <QGuiApplication>
 #include <QQmlApplicationEngine>
 #include <QIcon>
-#include <QtGlobal>
-#include <QCursor>
 
-#include "connectionsettings.h"
+#include "appStartup.hpp"
 
 void setCursorVisibility();
 
@@ -22,11 +20,10 @@ int main(int argc, char *argv[])
 
     QObject::connect(&engine, &QQmlApplicationEngine::quit, &QGuiApplication::quit);
 
-    setCursorVisibility();
+    QtPiDeck::setCursorVisibility();
+    QtPiDeck::registerQmlTypes();
 
-    // register all types in separate function
-    qmlRegisterType<ConnectionSettings>("data.connectionsettings", 1, 0, "ConnectionSettings");
-    const QUrl url(QStringLiteral("qrc:/main.qml"));
+    const QUrl url(QStringLiteral("qrc:/qml/pages/main.qml"));
     QObject::connect(&engine, &QQmlApplicationEngine::objectCreated,
                      &app, [url](QObject *obj, const QUrl &objUrl) {
         if (!obj && url == objUrl)
@@ -35,23 +32,4 @@ int main(int argc, char *argv[])
     engine.load(url);
 
     return app.exec();
-}
-
-void setCursorVisibility() {
-    auto hideCursorVar = qgetenv("HIDE_CURSOR");
-    auto hideCursor = [](){
-        QCursor cursor(Qt::BlankCursor);
-        QGuiApplication::setOverrideCursor(cursor);
-        QGuiApplication::changeOverrideCursor(cursor);
-    };
-    if (hideCursorVar.isEmpty()) {
-#ifdef Q_PROCESSOR_ARM
-        hideCursor();
-#endif
-        return;
-    }
-
-    if (hideCursorVar.toInt() != 0) {
-        hideCursor();
-    }
 }
